Refactor tests under xwe
