+++ tests/xwe/core/context/test_context_compressor.py
