+++ tests/xwe/core/nlp/test_deepseek.py
