+++ tests/xwe/core/nlp/test_nlp_processor.py
