+++ tests/xwe/core/output/test_output_manager.py
