+++ tests/xwe/core/state/test_state_manager.py
