+++ tests/xwe/core/test_attack_command.py
