+++ tests/xwe/core/test_attribute_buffs.py
