+++ tests/xwe/features/test_character_system.py
